BackEnd/requirements.txt
torch>=1.9.0
torch-geometric>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
numpy>=1.19.0
pandas>=1.3.0
pymysql>=1.0.2
pydantic>=1.8.0
scikit-learn>=0.24.0
python-dotenv>=0.19.0